Adding Game History Feature (Backend)

+++ src/main/java/com/github/floriangubler/eilemitweile/controller/MemberController.java
@@ -12,15 +12,17 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 @RestController
 @RequestMapping("/api/members")
-@Tag(name = "Members", description = "Coworkspace Members management endpoints")
+@Tag(name = "Members", description = "EileMitWeile Members management endpoints")
 public class MemberController {
 
     private final MemberService memberService;
@@ -30,39 +32,32 @@ public class MemberController {
     }
 
     @Operation(
-            summary = "Get Members",
-            description = "Get all Members (Only Admin)",
-            security = {@SecurityRequirement(name = "JWT Auth")}
-    )
-    @GetMapping
-    @PreAuthorize("hasRole('ADMIN')")
-    List<MemberEntity> loadUserBookings() {
-        return memberService.getMembers();
-    }
-
-    @Operation(
             summary = "Update a Member",
             description = "Admin updates a Member",
             security = {@SecurityRequirement(name = "JWT Auth")}
     )
     @PutMapping("/{memberid}")
-    @PreAuthorize("hasRole('ADMIN')")
     ResponseEntity<Void> updatemember (
             @io.swagger.v3.oas.annotations.parameters.RequestBody(description = "Member Update", required = true)
             @RequestBody(required = true)
             MemberDTO memberDTO,
             @Parameter(description = "MemberID", required = true)
             @PathVariable(name="memberid", required = true)
-            UUID memberid) {
+            UUID memberid,
+            Authentication authentication) {
         try{
-            String passwordHash = BCrypt.hashpw(memberDTO.getPassword(), BCrypt.gensalt());
-            memberService.update(new MemberEntity(UUID.randomUUID(), memberDTO.getEmail(), memberDTO.getFirstname(), memberDTO.getLastname(), passwordHash), memberid);
+            if(Objects.equals(UUID.fromString(authentication.getName()), memberid)){
+                String passwordHash = BCrypt.hashpw(memberDTO.getPassword(), BCrypt.gensalt());
+                memberService.update(new MemberEntity(UUID.randomUUID(), memberDTO.getEmail(), memberDTO.getFirstname(), memberDTO.getLastname(), passwordHash), memberid);
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else{
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
         } catch(UserNotFoundException e){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         } catch (UserAlreadyExistsException e){
             return new ResponseEntity<>(HttpStatus.CONFLICT);
         }
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @Operation(
@@ -71,14 +66,18 @@ public class MemberController {
             security = {@SecurityRequirement(name = "JWT Auth")}
     )
     @DeleteMapping("/{memberid}")
-    @PreAuthorize("hasRole('ADMIN')")
     ResponseEntity<Void> deletemember(
             @Parameter(description = "MemberID", required = true)
             @PathVariable(name = "memberid", required = true)
-            UUID memberid) {
+            UUID memberid,
+            Authentication authentication) {
             try{
-                memberService.delete(memberid);
-                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+                if(Objects.equals(UUID.fromString(authentication.getName()), memberid)){
+                    memberService.delete(memberid);
+                    return new ResponseEntity<>(HttpStatus.OK);
+                } else{
+                    return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+                }
             } catch(UserNotFoundException e){
                 return new ResponseEntity<>(HttpStatus.NOT_FOUND);
             }
